Updating README.md and adding tests cases

+++ geospatial_api/resources/geometry.py
@@ -9,10 +9,7 @@ from flask_smorest import Blueprint, abort
 
 from sqlalchemy import func
 
-from flask_jwt_extended import jwt_required
-
-from sqlalchemy.exc import IntegrityError
-from werkzeug.exceptions import BadRequest, Unauthorized
+from werkzeug.exceptions import BadRequest, Unauthorized, UnsupportedMediaType
 
 # custom libraries
 from geospatial_api.models.db import db
@@ -54,21 +51,28 @@ class GeometryResource(MethodView):
 
         return True
 
-    @jwt_required()
+
     def post(self) -> dict: 
         """
-            Creates a new geometry based on the description and geom parameters.
+            Creates a new geometry based on the description and geom parameters extracted from
+            the JSON payload of the request.
 
-        Args
-        ----
-            description: str,
-                Description of the geometry
-            geom: dict,
-                Geometry in GeoJSON format
+            The method expects a JSON payload with 'description' and 'geom' keys. It validates the inputs
+            and then inserts a new geometry into the database.
 
         Returns
         -------
-            A message indicating that the geometry was successfully added.
+            dict
+                A dictionary with a success message.
+
+        Raises
+        ------
+            ValueError
+                If 'description' or 'geom' are not provided or if 'geom' is not a valid GeoJSON object.
+            BadRequest
+                If the JSON payload is empty or malformed.
+            Exception
+                For any other server-side errors.
         """
         try:
 
@@ -79,47 +83,66 @@ class GeometryResource(MethodView):
 
             # Verifica se os parâmetros description e geom foram fornecidos e se
             # o geom é um objeto GeoJSON
-            if self.__validate_parameters(description=description, geom=geom):
-                geom = GeometryModel(
-                    description=description,
-                    geom=func.ST_GeomFromGeoJSON(json.dumps(geom)) 
-                )
+            if not description or not geom:
+                raise ValueError("Please provide description or geom")
+            geom = GeometryModel(
+                description=description,
+                geom=func.ST_GeomFromGeoJSON(json.dumps(geom)) 
+            )
 
-                db.session.add(geom)
-                db.session.commit()
-                db.session.close()
+            db.session.add(geom)
+            db.session.commit()
+            db.session.close()
 
-                return {"Success": f"Geometry added!"}, 201
+            return {"Success": f"Geometry added!"}, 201
         except ValueError as ve:
             abort(400, message=str(ve))
         except BadRequest as bre:
-            abort(400,
-                message="JSON file cannot be empty, must have description and geom!")
+            message = "JSON file cannot be empty, must have description and geom!"
+            abort(400, message=message)
+        except LookupError as le:
+            abort(404, message=str(le))
+        except UnsupportedMediaType as ume:
+            abort(415, message=str(ume))
         except Exception as e:
             abort(500, message=f"An error has occurred: {str(e)}")
 
-    @jwt_required()
+
     def get(self) -> dict:
         """
-            Returns a list of geometries based on the description and geom parameters.
+            Retrieves a list of geometries from the database.
 
-        Args
-        ----
-            description: str,
-                Description of the geometry
-            geom: dict,
-                Geometry in GeoJSON format
+            This method can retrieve a specific geometry by its ID, passed as a query parameter,
+            or a list of geometries that match the provided filters ('description' and 'geom') in the
+            JSON body of the request.
+
+            If no ID is provided, the method filters geometries based on the optional 'description'
+            and 'geom' fields. The 'geom' field should be in GeoJSON format.
 
         Returns
         -------
-            A list of geometries that match the description and geom parameters.
+            dict
+                A list of geometries matching the criteria, or an error message if no geometries are found.
+
+        Raises
+        ------
+            ValueError
+                If the 'description' and 'geom' parameters are missing or invalid.
+            BadRequest
+                If the request body is empty or not a valid JSON.
+            LookupError
+                If no geometry is found with the given ID or matching the filters.
+            UnsupportedMediaType
+                If the request content type is not supported.
+            Exception
+                For any other server-side errors.
         """
         try:
             id = request.args.get('id')
 
             # Busca pelo ID
             if id:
-                geometry = GeometryModel.query.get(id)
+                geometry = db.session.get(GeometryModel, id)
                 if not geometry:
                     raise LookupError(f"No geometry found with id {id}")
                 return [geometry.as_dict()]
@@ -135,12 +158,13 @@ class GeometryResource(MethodView):
             if not self.__validate_parameters(description=description, geom=geom):
                 raise ValueError("Invalid parameters: description and geom are required.")
 
-            geometry = GeometryModel.query
+            geometry = db.session.query(GeometryModel)
 
             if geom:
                 geom_json = json.dumps(geom)
                 geom = func.ST_GeomFromGeoJSON(geom_json)
                 geometry = geometry.filter(func.ST_Contains(GeometryModel.geom, geom))
+                #geometry = geometry.filter(func.ST_Contains(GeometryModel.geom, geom))
 
             if description:
                 geometry = geometry.filter_by(description=description)
@@ -158,24 +182,37 @@ class GeometryResource(MethodView):
             abort(400, message=message)
         except LookupError as le:
             abort(404, message=str(le))
+        except UnsupportedMediaType as ume:
+            abort(415, message=str(ume))
         except Exception as e:
             abort(500, message=f"An error has occurred: {str(e)}")
 
-    @jwt_required()
+
     def put(self) -> dict:
         """
-            Creates a new geometry based on the description and geom parameters.
+            Updates an existing geometry based on the provided ID.
 
-        Args
-        ----
-            description: str,
-                Description of the geometry
-            geom: dict,
-                Geometry in GeoJSON format
+            This method updates the 'description' and/or 'geom' fields of a geometry 
+            identified by the given ID, which is passed as a query parameter. 
+            The 'geom' field should be in GeoJSON format.
+
+            If the ID is not found in the database, a 404 error is returned.
 
         Returns
         -------
-            A message indicating that the geometry was successfully updated.
+            dict
+                A message indicating that the geometry was successfully updated.
+
+        Raises
+        ------
+            ValueError
+                If the request body is empty or the data is invalid.
+            BadRequest
+                If the ID parameter is missing or the JSON data is invalid.
+            LookupError
+                If no geometry is found with the given ID.
+            Exception
+                For any other server-side errors.
         """
         try:
             data = request.get_json()
@@ -191,7 +228,7 @@ class GeometryResource(MethodView):
             new_description = data.get("new_description", "")
             new_geom = data.get("new_geom", "")
 
-            geometry = GeometryModel.query.get(id)
+            geometry = db.session.get(GeometryModel, id)
             if not geometry:
                 raise LookupError(f"No geometry with id {id}")
 
@@ -202,7 +239,7 @@ class GeometryResource(MethodView):
 
             db.session.commit()
 
-            return {"Success": "The geometry was updated successfully"}, 201
+            return {"Success": "The geometry was updated successfully"}, 200
         except ValueError as ve:
             abort(400, message=str(ve))
         except BadRequest as bre:
@@ -213,19 +250,27 @@ class GeometryResource(MethodView):
         except Exception as e:
             abort(500, message=f"An error has occurred: {str(e)}")
 
-    @jwt_required()
+
     def delete(self) -> dict:
         """
-            Deletes a geometry based on the id parameter.
+            Deletes a geometry from the database based on the provided ID.
 
-        Args
-        ----
-            id: int,
-                ID of the geometry
+            This method deletes a geometry identified by the given ID, which is passed as 
+            a query parameter. If the ID is not found, a 404 error is returned.
 
         Returns
         -------
-            A message indicating that the geometry was successfully deleted.
+            dict
+                A message indicating that the geometry was successfully deleted.
+
+        Raises
+        ------
+            BadRequest
+                If the ID parameter is missing.
+            LookupError
+                If no geometry is found with the given ID.
+            Exception
+                For any other server-side errors.
         """
         try:
             id = request.args.get('id')
@@ -234,7 +279,10 @@ class GeometryResource(MethodView):
             if not id:
                 abort(400, message="Please provide an id")
 
-            geometry = GeometryModel.query.get(id)
+            geometry = db.session.get(GeometryModel, id)
+
+            if not geometry:
+                raise LookupError(f"No geometry found with id {id}")
 
             db.session.delete(geometry)
             db.session.commit()
@@ -246,5 +294,8 @@ class GeometryResource(MethodView):
         except BadRequest as bre:
             message = "JSON file cannot be empty, must have description and geom!"
             abort(400, message=message)
+        except LookupError as le:
+            abort(404, message=str(le))
         except Exception as e:
             abort(500, message=f"An error has occurred: {str(e)}")
+
